backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
greenlet==3.4.0
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
javalang==0.13.0
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pluggy==1.6.0
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rich==15.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
wheel==0.46.3
wrapt==2.1.2
